# Some simplification of the logic of the write operations