Here i try to reduce amount of code which should be read to understand write operations