@syndtr 2018-11-26
This rather complex because we want to try to merge concurrent `Put` and `Delete` into single batch. This PR pretty much undo that.