@pzhin 2018-11-26
When we try to put or delete single record, we wrap it into batch that can be merged to current write operation. https://github.com/syndtr/goleveldb/pull/246/commits/4469f287c68625e6fe21c7e40e0b7fe4d8f85551
It's looks like logic didn't change.