@pzhin 2018-11-26
Sorry, i found where is my mistake. And I wrote benchmark for parallel putting and it shows that my commits increase allocations.
```
$ benchstat old new                                                                                       
name             old time/op    new time/op    delta
DBPutParallel-4    7.69µs ± 6%    7.92µs ± 6%      ~     (p=0.075 n=10+10)

name             old speed      new speed      delta
DBPutParallel-4  15.1MB/s ± 5%  14.6MB/s ± 6%      ~     (p=0.075 n=10+10)

name             old alloc/op   new alloc/op   delta
DBPutParallel-4     37.9B ± 5%     76.4B ± 2%  +101.72%  (p=0.000 n=8+10)

name             old allocs/op  new allocs/op  delta
DBPutParallel-4      0.00           1.00 ± 0%     +Inf%  (p=0.000 n=10+10)
```